requirements.txt
fastapi
pydantic
python-telegram-bot==13.7
pyyaml
uvicorn
requests
